Enhanced MR Cleaning
Since the hard thresholding was problematic, and turning it off was
problematic too, I tried to find another solution to distinguish
negative and positive artifacts. Ventures into clustering were
unsuccessful. I finally decided to also use the previous frame, to help
decide which of fr_even and fr_odd was contaminated by an artifact. This
should be general enough.