Progressive input; run pupilometry only
The code now supports progressive video, not just interlaced. This is
because compressing an interlaced video leads to artifacts in both
fr_even and fr_odd, difficult to correct. I now read the original video
with the mmreader toolbox (could add it to Utils?), deinterlace it in
Matlab, and reencode it as a progressive .mp4.

Note: I also added the option to run the pupilometry only (no manual
input for calibration). That's just if I want to start the process and
leave my computer, I know it will complete the pupilometry and I can
then do the calibration.