This is a graphical interface to cut eyetracking videos into CAL/VAL/GAZE videos. Works on Windows and Mac, not tested yet on Linux.
The program is design to work on interlaced (but also works on deinterlaced) .mp4 videos. Each video should be contained in its own folder.

The gui allows you to select the start/end of calibration and validation, as well as find the first scanner artefact. When you press save, a prepare.mat file is written in rootfolder/name_of_video1/

Then, the function write_videos(videofile) reads the prepare.mat file and writes the CAL/VAL/GAZE videos, cutting up the original videofile.